requirements/production.txt
# Install everything that needs to be compiled
-r compiled.txt

# Django stuff
Django>=1.8

# Templates
django_compressor

# Security
bleach
#cef

# Logging
#raven

# L10n/i18n
#Babel
#-e git://github.com/clouserw/tower.git#egg=tower

# Caching
python-memcached

# Admin
django-debug-toolbar

# Feature testing
#django-waffle

# rest framework
# http://www.django-rest-framework.org/#installation
djangorestframework
django-filter

# REST Framework Authentication
django-allauth
django-rest-auth

# Position storing
django-geoposition

# cross origin requests
django-cors-headers
